Fix broken file links (#389)
Hi Christine,

this post gave me a lot and I want to give you something back by fixing some broken links to your GitHub files.

I chose to pin them to commits that were the latest right before you first published this post, so they are consistent with the snippets inside the post. I also pinned the gruvbox-css `default.nix` even though the link is working because this not only makes sure the contents behind the link will stay persistent with the post in case you choose to change something in there, but also because this enables others to clone the repo, check out the exact commit and try it out on their own.

I intentionally left some links (like those to your fish / spacemacs / etc. configs) unchanged so that they will always point to the latest files because they are just further examples of what you did and are not closely related to the post.

Thanks for the great explanations :heart: