Fix 2>&1 redirects
2>&1 comes *after* 1>, not before, if you want to combine outputs.

The way this works internally in bash is that when you do 2>&1, bash copies fd 1 and pastes it on fd 2. When you later change fd 1 to the file, fd 2 is still stdout.
But if you put the fd 1 redirect before, then fd 1 is changed first and fd 2 copies this change.