dhall/jobHistory: Fix endDate of "Archmage of Infrastructure"
The endDate of the position as Archmage of Infrastructure at Tailscale
is only one month after the start date, but daysWorked indicates that
the position lasted for more than one year. Likely, the year in endDate
is a typo.

This commit changes the endDate of the position to 2023, so that
`startDate + daysWorked = endDate' is satisfied.

Signed-off-by: Matthias Kruk <matthias.kruk@miraclelinux.com>