internal/lume: make pages for post series
I'm probably reading something wrong in the documentation for Lume, but
I wasn't able to find an obvious way to generate pages for each of the post
series programmatically.

This is a bit of a hacky workaround that makes the Lume driver write
individual JSX files for each series defined in the Dhall source. It
should work for my needs though.

Signed-off-by: Xe Iaso <me@xeiaso.net>