internal/lume: maintain an existing checkout of xesite
Okay so one of the main problems with this existing setup is that it
ended up recloning the GitHub repo every time the app booted. This made
startup slow by however long it takes for GitHub to serve the repo to
the client.

This change intelligently reuses the existing local clone of the git
repo so that it doesn't have to reclone every time.

There is a caveat for this though, it still makes you have an extra
clone of the repo in development mode. I'm going to accept this for now.

Signed-off-by: Xe Iaso <me@xeiaso.net>